group_1_intron_exon_classifier/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-doc
anyio
attrs
biopython
cbor2
certifi
charset-normalizer
click
einops
filelock
frozenlist
fsspec
grpclib
h11
h2
hf-xet
hpack
httpcore
httpx
huggingface-hub
hyperframe
idna
jinja2
joblib
markdown-it-py
markupsafe
mdurl
modal
mpmath
multidict
networkx
numpy
packaging
pandas
propcache
protobuf
pygments
python-dateutil
pyyaml
regex
requests
rich
safetensors
scikit-learn
scipy
setuptools
shellingham
six
sympy
synchronicity
threadpoolctl
tokenizers
toml
torch
tqdm
transformers
typer
typer-slim
types-certifi
types-toml
typing-extensions
urllib3
watchfiles
yarl
